+++ src/main.ts
@@ -53,6 +53,8 @@ export { default as ListAvatars } from "@/components/DataDisplay/ListAvatar";
 export * from "@/components/DataDisplay/ListAvatar";
 export { default as AutocompleteFilter } from "@/components/Inputs/AutocompleteFilter";
 export * from "@/components/Inputs/AutocompleteFilter";
+export { default as FileViewer } from "@/components/DataDisplay/FileViewer";
+export * from "@/components/DataDisplay/FileViewer";
 
 // Context
 export { default as ThemeProvider } from "@/context/Theme/ThemeProvider";