+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 # [Versions](https://github.com/Tracktor/design-system/releases)
 
-## v2.6.4
-- **[fix]** - fix global localization `fr` from `ThemeProvider`
+## v2.7.0
+- **[fix]** - Allow display of PDF files in the TimeLine component.
+- **[feat]** - Changed the image prop in TimeLineItem to file to support multiple file types.
+- **[feat]** - Added FileViewer component in the DataDisplay category.